refactor: move TagLabelPrefix to shared constants module BED-8743
